pyproject.toml
[project]
name = "agent-implementation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autopep8>=2.3.2",
    "docx2txt>=0.9",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.7",
    "python-dotenv>=1.2.1",
]
